Update README.md
The project is a simple Pong game created using Python as the backend language. Pong is a classic arcade game where two players control paddles on opposite sides of the screen, trying to hit a ball back and forth without letting it pass their paddle.The project demonstrates the fundamental concepts of game development using Python as the backend language. By utilizing Python's flexibility and ease of use, the game provides an interactive and enjoyable experience for players of all ages. Additionally, the project can serve as a starting point for further enhancements, such as adding more complex features or creating different game variations.